requirements.txt
backtrader==1.9.76.123
pandas==1.3.5
quantstats==0.0.47
pyyaml==6.0
black==21.12b0
flake8==4.0.1